@@ -39,7 +39,7 @@ export class LoginComponent implements OnInit {
       password: ['', Validators.required],
     });
 
-    this.returnUrl = this.route.snapshot.queryParams['returnUrl'] || 'dashboard';
+    this.returnUrl = this.route.snapshot.queryParams.returnUrl || 'dashboard';
   }
 
   get f() {
