@@ -1,6 +1,4 @@
-﻿
-
-export class User {
+﻿export class User {
   id: number;
   username: string;
   password: string;
